🏗️ test: add test client for Protocol V6
