+++ outscale/framework_provider.go
@@ -8,14 +8,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/provider"
 	"github.com/hashicorp/terraform-plugin-framework/provider/schema"
-	"github.com/hashicorp/terraform-plugin-framework/providerserver"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/types"
-	"github.com/hashicorp/terraform-plugin-go/tfprotov6"
-	"github.com/hashicorp/terraform-plugin-mux/tf5to6server"
-	"github.com/hashicorp/terraform-plugin-mux/tf6muxserver"
-	sdkresource "github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/outscale/terraform-provider-outscale/version"
 )
 
 var (
@@ -23,14 +17,22 @@ var (
 	_ provider.ProviderWithEphemeralResources = &frameworkProvider{}
 )
 
+type frameworkProvider struct {
+	version     string
+	onConfigure func(*OutscaleClientFW)
+}
+
 func New(version string) provider.Provider {
 	return &frameworkProvider{
 		version: version,
 	}
 }
 
-type frameworkProvider struct {
-	version string
+func NewWithConfigure(version string, on func(*OutscaleClientFW)) provider.Provider {
+	return &frameworkProvider{
+		version:     version,
+		onConfigure: on,
+	}
 }
 
 type ProviderModel struct {
@@ -198,6 +200,10 @@ func (p *frameworkProvider) Configure(ctx context.Context, req provider.Configur
 	resp.DataSourceData = *client
 	resp.ResourceData = *client
 	resp.EphemeralResourceData = *client
+
+	if p.onConfigure != nil {
+		p.onConfigure(client)
+	}
 }
 
 func (p *frameworkProvider) DataSources(ctx context.Context) []func() datasource.DataSource {
@@ -238,56 +244,5 @@ func (p *frameworkProvider) Resources(ctx context.Context) []func() resource.Res
 func (p *frameworkProvider) EphemeralResources(_ context.Context) []func() ephemeral.EphemeralResource {
 	return []func() ephemeral.EphemeralResource{
 		NewKeypairEphemeralResource,
-	}
-}
-
-func DefineTestProviderFactoriesV6() map[string]func() (tfprotov6.ProviderServer, error) {
-	return map[string]func() (tfprotov6.ProviderServer, error){
-		"outscale": func() (tfprotov6.ProviderServer, error) {
-			ctx := context.Background()
-			upgradedSdkServer, err := tf5to6server.UpgradeServer(
-				ctx,
-				Provider().GRPCProvider,
-			)
-
-			if err != nil {
-				return nil, err
-			}
-
-			providers := []func() tfprotov6.ProviderServer{
-				providerserver.NewProtocol6(New(version.GetVersion())),
-				func() tfprotov6.ProviderServer {
-					return upgradedSdkServer
-				},
-			}
-
-			muxServer, err := tf6muxserver.NewMuxServer(ctx, providers...)
-
-			if err != nil {
-				return nil, err
-			}
-
-			return muxServer.ProviderServer(), nil
-		},
-	}
-}
-
-func FrameworkMigrationTestSteps(sdkVersion string, config string) []sdkresource.TestStep {
-	return []sdkresource.TestStep{
-		{
-			ExternalProviders: map[string]sdkresource.ExternalProvider{
-				"outscale": {
-					VersionConstraint: sdkVersion,
-					Source:            "outscale/outscale",
-				},
-			},
-			Config: config,
-		},
-		{
-			ProtoV6ProviderFactories: DefineTestProviderFactoriesV6(),
-			Config:                   config,
-			PlanOnly:                 true,
-			ExpectNonEmptyPlan:       false,
-		},
 	}
 }
